# lib/api-zod/src/generated/types/imageSlot.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ImageSlot {
  /** The image slot key. */
  slot: string;
  /** The object path for the image. */
  objectPath: string;
  uploadedAt?: Date;
}