# lib/api-zod/src/generated/types/setImageSlotRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SetImageSlotRequest {
  /** The object path returned from the upload endpoint. */
  objectPath: string;
}